Affected services
- city_pass

Definition of Done:
- [ ] API is backwards compatible (`make openapi-diff`)
- [ ] Swagger UI up-to-date & tested (`make dev`)
- [ ] Dependencies updated (`make requirements`)
- [ ] Unittest coverage >90% (`make coverage`)
- [ ] Infrastructure config updated (aapp_azure_infra)
- [ ] Loadtests for relevant endpoints (aapp_testing_loadtests)

If manually deployed:
- [ ] Manual check on dev (o) & test (t), e.g. use feature via Swagger or admin panel
- [ ] Sanity check on test app

Fix status 500 responses

+++ city_pass/views/session_views.py
@@ -45,8 +45,9 @@ class SessionInitView(DeviceIdMixin, generics.CreateAPIView):
         Returns:
             Tuple[models.AccessToken, models.RefreshToken]: a new token pair
         """
-        old_sessions = models.Session.objects.filter(device_id=self.device_id).all()
-        old_sessions.delete()  # Remove old sessions for this device_id
+        if self.device_id:
+            old_sessions = models.Session.objects.filter(device_id=self.device_id).all()
+            old_sessions.delete()  # Remove old sessions for this device_id
 
         new_session = models.Session(device_id=self.device_id)
         new_session.save()
